Added ability to create xdripjs.json if not present and added validations to transmitter id

+++ bin/g5-transmitter.sh
@@ -8,24 +8,42 @@ function usage
   exit
 }
 
+function  generate_config
+{
+  echo "{\"transmitter_id\":\"${txId}\",\"sensor_id\":\"\",\"mode\":\"not-expired\",\"pump_units\":\"mg/dl\",\"fake_meter_id\":\"000000\"}" | jq . > $config
+}
+
 numArgs=$#
 if [ $(bc <<< "$numArgs < 1") -eq 1 ] || [ $(bc <<< "$numArgs > 1") -eq 1 ]; then
-# use -m or -c to specify minutes ago and/or g6 code
   usage
   exit
 fi
 
 txId=$1
-echo ${txId}
+
+#validate transmitter id
+if [ ${#txId} -ne 6 ]; then
+  echo "Invalid transmitter id; should be 6 characters long"
+  exit
+elif [ ${txId:0:1} == 8 ]; then
+  echo "setting G6 transmitter: $txId"
+else
+  echo "setting G5 transmitter: $txId"
+fi
 
 #Update config file
 config="/root/myopenaps/xdripjs.json"
-tmp=$(mktemp)
-jq --arg txId "$txId" '.transmitter_id = $txId' "$config" > "$tmp" && mv "$tmp" "$config"
+if [ -e "$config" ]; then
+  echo "config found, updating xdripjs.json"
+  tmp=$(mktemp)
+  jq --arg txId "$txId" '.transmitter_id = $txId' "$config" > "$tmp" && mv "$tmp" "$config"
+else
+  echo "Config file does not exist, generating xdripjs.json"
+  generate_config
+fi
 
 #Post message
 MESSAGE="${HOME}/myopenaps/monitor/xdripjs/cgm-transmitter.json"
 epochdate=$(date +'%s%3N')
-
 echo "[{\"date\":\"${epochdate}\",\"type\":\"New Transmitter\",\"txId\":\"${txId}\"}]" >  $MESSAGE
 cat $MESSAGE
